requirements.txt
Flask==2.3.3
requests==2.31.0
beautifulsoup4==4.13.0
PyYAML==6.0.1
jieba==0.42.1
python-Levenshtein==0.21.1
googletrans==4.0.0-rc1
flask-cors==4.0.0